create web socket test data and move awaitedsockettasktest
