+++ tests/Nakama.Tests/Socket/WebSocketTestData.cs
@@ -1,0 +1,34 @@
+/**
+* Copyright 2020 The Nakama Authors
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+* http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*/
+
+using System.Collections;
+using System.Collections.Generic;
+
+namespace Nakama.Tests.Socket
+{
+    public delegate WebSocketAdapter TestAdapterFactory();
+
+    public class WebSocketTestData : IEnumerable<object[]>
+    {
+        IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
+
+        public IEnumerator<object[]> GetEnumerator()
+        {
+            TestAdapterFactory textAdapterFactory = () => new WebSocketAdapter();
+            yield return new object[]{textAdapterFactory};
+        }
+    }
+}
